requirements/requirements.txt
appdirs==1.4.2
appnope==0.1.0
decorator==4.0.11
Django==1.9.9
django-debug-toolbar==1.6
django-extensions==1.7.7
django-filter==1.0.1
djangorestframework==3.5.4
ipython==5.3.0
ipython-genutils==0.1.0
Markdown==2.6.8
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.2
wcwidth==0.1.7
